👌 Spliting badge components
